requirements.txt
fastapi>=0.109.1,<0.111
uvicorn[standard]>=0.27.0,<0.31
sqlalchemy>=2.0.23,<2.1
psycopg2-binary>=2.9.10,<3.0
python-dotenv>=1.0.0,<2.0
pydantic[email]>=2.5.0,<3.0
email-validator>=2.0.0,<3.0
bcrypt>=4.0.1,<5.0
httpx>=0.27.0,<0.28
uliweb-alembic